Add menu and fixed issue with installing software
Added a menu bar at the top that allows for importing and exporting settings. It also includes the following buttons: About, link to GitHub, link to report issues, and a help button to guide the user if needed. This starts some ground work for issue #15 and issue #5.

Also added in a fix for when trying to install applications that had a space in them. Fixes issue #16.